refactor: 함수 기능 별로 파일 분리하여 작성
